Trigger Queue now locks on execution
Can add multibly items to trigger queue that both will result in
true and false and all items will exdcute in a circular queue

+++ triggers_test.go
@@ -20,7 +20,7 @@ func TestTemp(t *testing.T) {
 	}
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			assert.Equal(t, Triggers[test.name](test.param), test.expected)
+			assert.Equal(t, Triggers[test.name](float64(test.param)), test.expected)
 		})
 	}
 }
@@ -39,7 +39,7 @@ func TestWeight(t *testing.T) {
 	}
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			assert.Equal(t, Triggers[test.name](test.param), test.expected)
+			assert.Equal(t, Triggers[test.name](float64(test.param)), test.expected)
 		})
 	}
 }
